rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
qdrant-client==1.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.7
tiktoken==0.5.2
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
